autogen.sh: Rewrite
The goal is twofold: firstly, we want to be able to make parts
of it into reusable modules, and secondly we'd like to reduce
the amount of duplicated code. Moreover, since we're making
heavy changes to the code anyway, we might as well make sure
it follows a somewhat consistent coding style too.

To reduce code duplication, we introduce a new --dry-run
option, which can be used by third parties to figure out
whether calling autogen.sh is necessary or not: this allows
us to get rid of the reimplementation of part of the logic
in cfg.mk and guarantee they'll never get out of sync.

In order to be able to move forward with modularization down
the line, we extract the gnulib-specific part of the submodule
logic into three functions which adhere to a fairly generic
API that's not tied to gnulib.

Other changes include: removing assumptions about gnulib being
the only submodule, both in actual logic and in messages to
the user; making dirty submodules checking and cleaning
entirely independent of whether or not updating submodules is
needed; removing the use of 'set -e' and handling errors
explicitly instead.
Message-Id: <1492525878-21819-2-git-send-email-abologna@redhat.com>