Introduce virStreamInData
This is just an internal API, that calls corresponding function
in stream driver. This function will set @data=1 if the
underlying file is in data section, or @data=0 if it is in a
hole. At any rate, @length is set to number of bytes remaining in
the section the file currently is.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <11e2793cb192de9d66f0147c6c8d9910d56a0428.1492682033.git.mprivozn@redhat.com>