virnetclientstream: Introduce virNetClientStreamHandleSkip
This is a function that handles an incoming STREAM_SKIP packet.
Even though it is not wired up yet, it will be soon. At the
beginning do couple of checks whether server plays nicely and
sent us a STREAM_SKIP packed only after we've enabled sparse
streams. Then decodes the message payload to see how big the hole
is and stores it in passed @length argument.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <14feb308a8e92929292718303a5a4ef6388d723e.1492682034.git.mprivozn@redhat.com>