remote_driver: Implement virStreamSkip
Now that we have RPC wrappers over VIR_NET_STREAM_SKIP we can
start wiring them up. This commit wires up situation when a
client wants to send a hole to daemon.

To keep stream offsets synchronous, upon successful call on the
daemon skip the same hole in local part of the stream.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <a613c7c273a17cfc60ee5bd629868c99a171a6e6.1492682034.git.mprivozn@redhat.com>