daemon: Don't call virStreamInData so often
While virStreamInData() should be a small and quick function, in
our implementation it seeks multiple times. Moreover, it is
called even if we know that we are in data. Well, quite. If we
track its return values and do some basic math with them, we can
end up calling virStreamInData right at the boundaries of a data
section or a hole and nowhere else.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <42ed53310b4f05787b0d76f700f48b29d3a982f6.1492682034.git.mprivozn@redhat.com>