virsh: Implement sparse stream to vol-download
The command grew new --sparse switch that does nothing more than
enables the sparse streams feature for this command. Among with
the switch new helper function is introduced: virshStreamSkip().
This is the callback that is called whenever daemon sends us a
hole. In the callback we reflect the hole in underlying file by
seeking as many bytes as told.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <a5074be2f86b5bafcf83368110f48cf8cc6f724c.1492682034.git.mprivozn@redhat.com>