secret: Use consistent naming for variables
When processing a virSecretPtr use 'secret' as a variable name.

When processing a virSecretObjPtr use 'obj' as a variable name.

When processing a virSecretDefPtr use 'def' as a variable name,
unless a distinction needs to be made with a 'newdef' such as
virSecretObjListAddLocked (which also used the VIR_STEAL_PTR macro
for the configFile and base64File).

NB: Also made a slight algorithm adjustment for virSecretObjListRemove
to check if the passed obj was NULL rather than having the caller
virSecretLoad need to check prior to calling.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170424180023.4695-6-jferlan@redhat.com>