qemu: don't kill qemu process on restart if networkNotify fails
Nothing that could happen during networkNotifyActualDevice() could
justify unceremoniously killing the qemu process, but that's what we
were doing.

In particular, new code added in commit 85bcc022 (first appearred in
libvirt-3.2.0) attempts to reattach tap devices to their assigned
bridge devices when libvirtd restarts (to make it easier to recover
from a restart of a libvirt network). But if the network has been
stopped and *not* restarted, the bridge device won't exist and
networkNotifyActualDevice() will fail.

This patch changes qemuProcessNotifyNets() to return void, so that
qemuProcessReconnect() will soldier on regardless of what happens (any
errors will still be logged, but we won't terminate the guest).

Partially resolves: https://bugzilla.redhat.com/1442700
Message-Id: <20170425163400.8140-2-laine@laine.org>