util: mdev: Use a local variable instead of an invalid pointer reference
virMediatedDeviceListAdd calls VIR_APPEND_ELEMENT which normally clears
the element to be added, thus the pointer must not be used afterwards,
but because the pointer here is passed by value, what gets cleared is
a copy of the original pointer that got created on the stack
automatically when calling the function. The fact that it works now is
just a coincidence and a bug in virMediatedDeviceListAdd that needs to
be fixed in a separate commit. Therefore, use a local variable to hold
data, rather than accessing the pointer later.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <f795415943ebebdad5ff2485dda0dc0ca4b2fe51.1493365484.git.eskultet@redhat.com>
Reviewed-by: Laine Stump <laine@laine.org>