qemu: change the name of tap device for a tap and bridge network
Creating tap device and adding the device to bridge are not atomic operation.
Similarly deleting tap device and removing it from bridge are not atomic operation.
The Problem occurs when two vms start and shutdown. When one vm with the nic
named "vnet0" stopping, it deleted tap device but not removing port from bridge.
At this time, another vm created the tap device named "vnet0" and added port to the
same bridge. Then, the first vm deleted the tap device from the same bridge.
Finally, the tap device of the second vm don't attached to the bridge.
So, we can add domid to vm's nic name. For example, the vm's domid is 1 and vnet0
is renamed to vnet1.0.

Signed-off-by: ZhiPeng Lu <lu.zhipeng@zte.com.cn>
Message-Id: <1493371399-25799-1-git-send-email-lu.zhipeng@zte.com.cn>