docs: Add details about panic device behavior
When a guest triggers the panic device, QEMU pauses the guest, and
libvirt takes the action specified by on_crash. This can interfere
with the guest's own crash handling actions (like writing a dump file
and rebooting itself) if the guest triggers the panic device first (as
Windows does via the Hyper-V crash CPU feature). This is not an
obvious side effect of a notification mechanism, so it's worth
explaining.

Signed-off-by: Ed Swierk <eswierk@skyportsystems.com>
Message-Id: <1493400914-6040-1-git-send-email-eswierk@skyportsystems.com>