qemu: replace nested job with interruptible async job state
Nested job is a construction that gives way to run regular jobs while async job
is running. But the period when another job is actually can be started is when
async job waits for some event with domain lock dropped. Let's code this
condition straitforward using asyncInterruptible flag. Upon awaking in async
job it is important to wait while any regular job that is started meanwhile is
finished so the two are not overlapped and don't use domain or monitor object
simultaneously.

This solution isolates jobs even better than nested job appoach. In the
latter it is possible that concurrent job awaits reply from monitor, then
monitor event is delivered and async job awakes and continues to run until it
needs to send command to monitor. Only at this moment it stops on acquiring
nested job condition.

qemuDomainObjEnterMonitorInternal needs asyncJob argument and can not use
domain job state because the function can be called simultaneously from async
job and concurrend job and can not determine context due to nesting without help
from callers. With qemuDomainObjWaitInternal there is no cases when this
function is called from concurrent job when async job is running currently. But
just for the case detection is possible and implemented because this state of
affairs can be changed one day.

So qemuDomainObj{Wait,Sleep} are safe to use from regular or async job context.
There will be no situations when qemuDomainObjEnterMonitor should be changed to
qemuDomainObjEnterMonitorAsync because function begins being used from async
job context.
Message-Id: <1493719596-35375-2-git-send-email-nshirokovskiy@virtuozzo.com>