qemu: backup: prepare backup destination
Prepare here is usual preparation for a disk to be used by qemu made
by qemuDomainDiskChainElementPrepare. That is set security labels,
add to lock manager and whitelist in cgroups. All three are related to
backup target too. Adding to a lock manager is less obvious but
can be useful if mirations with a running backup will be possible.

Also create file in case of backup to file if it is not exist in
order to set securitly labels etc.
Message-Id: <1493965351-117429-10-git-send-email-nshirokovskiy@virtuozzo.com>