qemu: reduce packet loss rate for vm with macvtap passthrough mode in migration
Before libvirt that calls virNetDevMacVLanCreateWithVPortProfile sets mac address
or vlan of a Virtual Function(VF) linked to a macvtap passthrough device of migration
destination host in migration start step. If we ping the migrating vm,
we get the network does not pass. Because VFs of migration source and destination
have the same MAC address. The patch later calling qemuMigrationVPAssociatePortProfiles
sets mac address of VF in migration finish step instead of start step.
The patch aims to reduce packet loss rate.

Signed-off-by: ZhiPeng Lu <lu.zhipeng@zte.com.cn>
Message-Id: <1494553604-15950-1-git-send-email-lu.zhipeng@zte.com.cn>