virnetclientstream: Introduce virNetClientStreamHandleHole
This is a function that handles an incoming STREAM_HOLE packet.
Even though it is not wired up yet, it will be soon. At the
beginning do couple of checks whether server plays nicely and
sent us a STREAM_HOLE packed only after we've enabled sparse
streams. Then decodes the message payload to see how big the hole
is and stores it in passed @length argument.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <9920c5b27004b45826f6458d340d2a4f08864dcb.1494943031.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>