remote_driver: Implement virStreamSendHole
Now that we have RPC wrappers over VIR_NET_STREAM_HOLE we can
start wiring them up. This commit wires up situation when a
client wants to send a hole to daemon.

To keep stream offsets synchronous, upon successful call on the
daemon skip the same hole in local part of the stream.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <bcc1baeee827cb6086f2efef5cc8b0c33f3758f0.1494943031.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>