fdstream: Implement sparse stream
Basically, what is needed here is to introduce new message type
for the messages passed between the event loop callbacks and the
worker thread that does all the I/O. The idea is that instead of
a queue of read buffers we will have a queue where "hole of size
X" messages appear. That way the even loop callbacks can just
check the head of the queue and see if the worker thread is in
data or a hole section and how long the section is.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <ab58d7419e394cae244e0e06d7b9082228067db3.1494943031.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>