virsh: Implement sparse stream to vol-download
The command grew new --sparse switch that does nothing more than
enables the sparse streams feature for this command. Among with
the switch new helper function is introduced: virshStreamSkip().
This is the callback that is called whenever daemon sends us a
hole. In the callback we reflect the hole in underlying file by
seeking as many bytes as told.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <e827945797681071f362333806bb7d59277ec70d.1494943031.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>