libxl: add default listen address for VNC and spice
If a VNC listen address is not specified in domXML, libxl
will default to 127.0.0.1, but this is never reflected in the domXML.
In the case of spice, a missing listen address resulted in listening
on all interfaces, i.e. '0.0.0.0'. If not specified, set the listen
address in virDomainGraphicsDef struct to the libxl default when
creating the frame buffer device. Additionally, set default spice
listen address to 127.0.0.1.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>
Message-Id: <20170519223341.5090-3-jfehlig@suse.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>