virStreamSparseSendAll: Reset @want in each iteration
There's a slight problem with the current function. Assume we are
currently in a data section and we have say 42 bytes until next
section (okay, in reality sections are capped to the sector size,
but bear with me). Therefore, just before (handler) is called to
fill up the buffer with data, @want is changed to 42 to match the
amount of data left in the current section. However, after hole
is processed, we are back in data section but with incredibly
small @want size. Nobody will ever reset it back. This results in
incredible data fragmentation.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <c6ecc5048fc12c036d705c64f42b9d4f23d1a1e9.1495451186.git.mprivozn@redhat.com>