interface: Introduce virInterfaceObjEndAPI
For now it'll just call the virInterfaceObjUnlock, but a future adjustment
will do something different.

The virInterfaceObjUnlock is now private to virinterfaceobj.c with a
short term forward reference.

Additionally, make virInterfaceObjLock private since it's only used in
virinterfaceobj anyway. For now this will involved creating a forward
reference, but this will go away soon too.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20170526115910.3977-10-jferlan@redhat.com>