virStream*All: Call virStreamAbort() more frequently
Our documentation to all four virStreamRecvAll, virStreamSendAll,
virStreamSparseRecvAll, virStreamSparseSendAll says that if these
functions fail, virStreamAbort() is called. But that is not
necessarily true. For instance all of these functions allocate a
buffer to work with. If the allocation fails, no virStreamAbort()
is called despite -1 being returned. It's the same story with
argument sanity checks and a lot of other checks.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <d32ffa9f4eca2b926c386fa2cfa3199a4e83e158.1496318427.git.mprivozn@redhat.com>