Make virNetClientStreamEventRemoveCallback's errors optional
There will be new place for this function to be called, but we need
not to report error from there.  Since the virNetClientStream
structure is private, there is no way to check for st->cb from outside
this file, so we need to make the error reporting optional.  The
function name is already long enough to have yet another Quiet suffix
like some other functions do, plus it is only called from two places,
so just add bool for that.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <43d414ff05bc75123cb67b89f47c4ef498acc500.1496318548.git.mkletzan@redhat.com>